Reorganize a few things
